backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
alembic==1.13.2
pytest==8.3.1
httpx==0.27.0
